[refactor] separated client and server/database to hae their own dependencies.  Everything still works =)
