+++ server/database/mySQL/helpers/tripHelper.js
