+++ server/database/mySQL/models/trip.js
