+++ server/example.config.js
@@ -1,0 +1,3 @@
+const API_KEY = "API_KEY_GOES_HERE";
+
+module.exports = API_KEY;
