+++ server/example.env
@@ -1,0 +1,2 @@
+USERNAME=your_mysql_username
+PASSWORD=your_mysql_password
