+++ server/googleApiController.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import API_KEY from '../config';
+import API_KEY from './config';
 
 const getPlaces = (req, res) => {
   let { query } = req.query;
