+++ server/test/auth.test.js
@@ -1,5 +1,5 @@
 import '@babel/polyfill';
-import app from '../server/index';
+import app from '../index';
 
 import { Trip, User } from '../database/mySQL/models/joinTable';
 import sql_connection from '../database/mySQL/index';
