+++ socket-server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "I-Kinerary",
+  "name": "I-Kinerary-socket",
   "version": "1.0.0",
   "main": "index.js",
   "repository": "https://github.com/omegak911/I-Kinerary.git",
